Move Ops to new syntax
We discussed the possibility to add side structure to propagate and compute the
symbolic expression. Our first attempt was to do that using the type system,
did not prove to be a good idea. Thus, we simplified the type system by
introducing a generic ‘term’ type and let the side data structures (still to be
added) to the heavy lifting.