Handle nested equationOp
We do still have some limitations. Specifically we require each
equationOp to have a single use. Situation like:

```
%0 = linnea.equation {
 ...
}

%1 = linnea.equation {
  use(%0)
}

print %1
print %0

```

are not allowed. The first equation has two uses. The first equation is pulled
inside the second and the expression is optimized symbolically thus we lose the
1:1 matching between the IR and the symbolic expression.  It become difficult
to understant what value need to replace the %0 of the first equation.